requirements.txt
anthropic>=0.40.0
openai>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.5.0
tenacity>=8.2.0
